# fix(dstest): Cache the reading of TestCases, speeding up unit tests